Move psalm entry-point to a Cli command class
